Revert "Revert "PlatformResourceType""
This reverts commit 03b767aaacd240dfb73ae5122a37817b7dea728c.
