+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/model/v2/ResourceType.java
@@ -7,19 +7,28 @@ import java.util.Map;
 import org.apache.commons.lang3.ObjectUtils;
 
 import com.sap.cloud.lm.sl.common.util.MapUtil;
+import com.sap.cloud.lm.sl.mta.model.ElementContext;
+import com.sap.cloud.lm.sl.mta.model.NamedElement;
 import com.sap.cloud.lm.sl.mta.model.ParametersContainer;
+import com.sap.cloud.lm.sl.mta.model.VisitableElement;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
 
-public class ResourceType extends com.sap.cloud.lm.sl.mta.model.v1.ResourceType implements ParametersContainer {
+public class ResourceType implements VisitableElement, NamedElement, ParametersContainer {
 
+    private String name;
+    private String resourceManager;
     private Map<String, Object> parameters;
 
     protected ResourceType() {
 
     }
 
-    @Override
-    public Map<String, Object> getProperties() {
-        throw new UnsupportedOperationException();
+    public String getName() {
+        return name;
+    }
+
+    public String getResourceManager() {
+        return resourceManager;
     }
 
     @Override
@@ -27,9 +36,12 @@ public class ResourceType extends com.sap.cloud.lm.sl.mta.model.v1.ResourceType 
         return MapUtil.unmodifiable(parameters);
     }
 
-    @Override
-    public void setProperties(Map<String, Object> properties) {
-        throw new UnsupportedOperationException();
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setResourceManager(String resourceManager) {
+        this.resourceManager = resourceManager;
     }
 
     @Override
@@ -37,11 +49,18 @@ public class ResourceType extends com.sap.cloud.lm.sl.mta.model.v1.ResourceType 
         this.parameters = new LinkedHashMap<>(parameters);
     }
 
-    public static class Builder extends com.sap.cloud.lm.sl.mta.model.v1.ResourceType.Builder {
+    @Override
+    public void accept(ElementContext context, Visitor visitor) {
+        visitor.visit(context, this);
+    }
 
+    public static class Builder {
+
+        protected String name;
+        protected String resourceManager;
+        protected Map<String, Object> properties;
         protected Map<String, Object> parameters;
 
-        @Override
         public ResourceType build() {
             ResourceType result = new ResourceType();
             result.setName(name);
@@ -50,13 +69,20 @@ public class ResourceType extends com.sap.cloud.lm.sl.mta.model.v1.ResourceType 
             return result;
         }
 
-        public void setParameters(Map<String, Object> parameters) {
-            this.parameters = parameters;
+        public void setName(String name) {
+            this.name = name;
         }
 
-        @Override
+        public void setResourceManager(String resourceManager) {
+            this.resourceManager = resourceManager;
+        }
+
         public void setProperties(Map<String, Object> properties) {
             throw new UnsupportedOperationException();
+        }
+
+        public void setParameters(Map<String, Object> parameters) {
+            this.parameters = parameters;
         }
 
     }
